backend-2-ai_engine_service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
numpy==1.26.4
pandas==2.2.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
scikit-learn==1.3.2
xgboost==1.7.5
redis==4.5.5
pydantic==1.10.9
python-dotenv==1.0.0
uvicorn==0.22.0
